requirements.txt
numpy>=1.24
scipy>=1.10
scikit-image>=0.22
opencv-python>=4.8
matplotlib>=3.7
tqdm>=4.66
Pillow>=10.0
patchify>=0.2.3

torch>=2.2
torchvision>=0.17

monai-weekly[nibabel,tqdm]>=1.4

staintools>=2.1.2
